requirements.txt
flask
flask-cors
requests
python-dotenv
SpeechRecognition
pyaudio
pyttsx3
PyPDF2
scikit-learn
groq
beautifulsoup4